Add plugin restore tests.
Also live restore is stable now. So move experimental tests out to stable.

(cherry picked from commit 006d58d7e6756848058bd69a12c698c44e64efd1)
